+++ test/MqttNetworkBuilder/MqttNetworkBuilderMqttBrokerBuilderConfiguration.h
@@ -13,16 +13,13 @@
 class MqttNetworkBuilderMqttBrokerBuilderConfiguration {
  public:
   const MqttBrokerTestType type;
-  const int32_t maxBrokerCount;
+  const uint16_t maxBrokerCount;
 
   const MqttBrokerProtocolTestType protocol_type; // example tcp
   const std::string brokerAddress; // example localhost
   const uint16_t minBrokerPort; // example 1883
-  MqttNetworkBuilderMqttBrokerBuilderConfiguration(const MqttBrokerTestType type,
-                                            const int32_t max_broker_count,
-                                            const MqttBrokerProtocolTestType protocol_type,
-                                            const std::string &broker_address,
-                                            const uint16_t min_broker_port);
+  MqttNetworkBuilderMqttBrokerBuilderConfiguration(const MqttBrokerTestType type, const uint16_t max_broker_count, const MqttBrokerProtocolTestType protocol_type, const std::string &broker_address,
+                                                   const uint16_t min_broker_port);
   static const MqttNetworkBuilderMqttBrokerBuilderConfiguration GetDefaultConfiguration();
 };
